+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres
 to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.1.1](https://github.com/unzerdev/php-sdk/compare/3.1.0..3.1.1)
+
+### Deprecated
+* `\UnzerSDK\Resources\Payment::getCancellation`, please use `getCancellation` method of `\UnzerSDK\Resources\TransactionTypes\Charge` or `\UnzerSDK\Resources\TransactionTypes\Authorization` instead, depending on your use-case.
+
 ## [3.1.0](https://github.com/unzerdev/php-sdk/compare/3.0.0..3.1.0)
 
 ### Added